test(backends): move json `to_pyarrow` test to work against all backends
